requirements.txt
asttokens==2.2.1
backcall==0.2.0
dataclasses-json==0.5.7
decorator==5.1.1
executing==1.2.0
ffmpeg-python==0.2.0
future==0.18.3
ipython==8.13.2
jedi==0.18.2
Jinja2==3.1.2
jsonpickle==3.0.1
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
networkx==3.1
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.38
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
pyvis==0.3.2
six==1.16.0
stack-data==0.6.2
traitlets==5.9.0
typing-inspect==0.9.0
typing_extensions==4.6.2
wcwidth==0.2.6
